remove forum and contact
