+++ src/components/Footer/Footer.tsx
@@ -1,22 +1,22 @@
-import React from "react"
-import { useHistory } from "react-router-dom"
-import { MDBCol, MDBContainer, MDBRow, MDBFooter } from "mdbreact"
-import { GiWalkingBoot } from "react-icons/gi"
-import axios, { AxiosError, AxiosResponse } from "axios"
-import { connect } from "react-redux"
-import { Dispatch, bindActionCreators } from "redux"
-import RootState from "../../types/State/Root/State"
-import Account from "../../types/PostRequests/Account"
+import React from "react";
+import { useHistory } from "react-router-dom";
+import { MDBCol, MDBContainer, MDBRow, MDBFooter } from "mdbreact";
+import { GiWalkingBoot } from "react-icons/gi";
+import axios, { AxiosError, AxiosResponse } from "axios";
+import { connect } from "react-redux";
+import { Dispatch, bindActionCreators } from "redux";
+import RootState from "../../types/State/Root/State";
+import Account from "../../types/PostRequests/Account";
 import {
   setLoggedIn,
   setUserFirstName,
   setUserId,
   setUserLastName,
-} from "../../actions/actions"
-import "./Footer.css"
-import FooterLink from "./FooterLink"
-import FooterCol from "./FooterCol"
-import * as Actions from "../../types/Actions"
+} from "../../actions/actions";
+import "./Footer.css";
+import FooterLink from "./FooterLink";
+import FooterCol from "./FooterCol";
+import * as Actions from "../../types/Actions";
 
 interface Props {
   loggedIn: boolean;
@@ -35,41 +35,41 @@ const Footer: React.FC<Props> = (props: Props) => {
     setUserId,
     setUserFirstName,
     setUserLastName,
-  } = props
+  } = props;
 
-  const pushSlug: Function = useHistory().push
+  const pushSlug: Function = useHistory().push;
 
   const logOut = () => {
-    localStorage.clear()
-    localStorage.setItem("popupVisible", "false")
-    setLoggedIn(false)
-    setUserId("")
-    setUserFirstName("")
-    setUserLastName("")
-    alert("Logged out successfully.")
-    pushSlug("/forum")
-  }
+    localStorage.clear();
+    localStorage.setItem("popupVisible", "false");
+    setLoggedIn(false);
+    setUserId("");
+    setUserFirstName("");
+    setUserLastName("");
+    alert("Logged out successfully.");
+    pushSlug("/forum");
+  };
 
   const deleteAccount = () => {
     const payload: Account = {
       userId: userId,
-    }
+    };
 
     axios
       .delete("https://city-walks-production.up.railway.app/delete-account", {
         data: payload,
       })
       .then((res: AxiosResponse) => {
-        console.log("account deleted")
+        console.log("account deleted");
       })
       .catch((err: AxiosError) => {
-        console.log(err, "account not deleted, try again.")
-      })
+        console.log(err, "account not deleted, try again.");
+      });
 
-    alert("Account Deleted. You can sign up again at any time.")
-    logOut()
-    window.scrollTo(0, 0)
-  }
+    alert("Account Deleted. You can sign up again at any time.");
+    logOut();
+    window.scrollTo(0, 0);
+  };
 
   return (
     <MDBFooter className='font-small pt-4 mt-4 footer'>
@@ -91,19 +91,12 @@ const Footer: React.FC<Props> = (props: Props) => {
               <FooterLink path='/walks' title='Walks' />
               <FooterLink path='/about' title='About' />
               <FooterLink path='/sights' title='Sights' />
-              <FooterLink path='/contact' title='Contact' />
-              <FooterLink path='/forum/login' title='User Login' />
             </FooterCol>
           </MDBCol>
           <MDBCol md='3' className='footer-extra-links'>
             <FooterCol title='Extra Links'>
               <FooterLink path='/admin' title='Admin Portal' />
               <FooterLink path='/privacy' title='Privacy Policy' />
-              {loggedIn && (
-                <li className='list-unstyled'>
-                  <a onClick={deleteAccount}>Delete Account</a>
-                </li>
-              )}
             </FooterCol>
           </MDBCol>
         </MDBRow>
@@ -121,13 +114,13 @@ const Footer: React.FC<Props> = (props: Props) => {
         </MDBContainer>
       </div>
     </MDBFooter>
-  )
-}
+  );
+};
 
 const mapStateToProps: (state: RootState) => void = (state) => ({
   loggedIn: state.loginState.loggedIn,
   userId: state.loginState.userId,
-})
+});
 
 const mapDispatchToProps = (dispatch: Dispatch) => {
   return {
@@ -136,7 +129,7 @@ const mapDispatchToProps = (dispatch: Dispatch) => {
       { setLoggedIn, setUserId, setUserFirstName, setUserLastName },
       dispatch
     ),
-  }
-}
+  };
+};
 
-export default connect(mapStateToProps, mapDispatchToProps)(Footer)
+export default connect(mapStateToProps, mapDispatchToProps)(Footer);
