--- src/components/Pages/Contact/Contact.css
@@ -1,33 +1,0 @@
-/** @format */
-.contact-grid {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  grid-gap: 30px;
-}
-@media (max-width: 767px) {
-  .contact-grid {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
-
-.contact-main-img img {
-  width: 100%;
-}
-
-.contact-form textarea {
-  min-height: 118px;
-}
-
-#contact-submit-btn {
-  float: right;
-}
-@media (max-width: 767px) {
-  #contact-submit-btn {
-    width: 100%;
-    padding: 10px;
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
-
-/*# sourceMappingURL=Contact.css.map */
