--- src/components/Pages/Contact/Contact.scss
@@ -1,34 +1,0 @@
-/** @format */
-
-.contact-grid {
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  grid-gap: 30px;
-
-  @media (max-width: 767px) {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
-
-.contact-main-img {
-  img {
-    width: 100%;
-  }
-}
-
-.contact-form {
-  textarea {
-    min-height: 118px;
-  }
-}
-
-#contact-submit-btn {
-  float: right;
-
-  @media (max-width: 767px) {
-    width: 100%;
-    padding: 10px;
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
