--- src/components/Pages/LoginPage/ForgotPasswordForm/ForgotPasswordForm.scss
@@ -1,11 +1,0 @@
-/** @format */
-
-.forgot-password-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.forgot-password-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
