--- src/components/Pages/LoginPage/LoggedInView/Boards.css
@@ -1,10 +1,0 @@
-/** @format */
-.boards-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-/*# sourceMappingURL=Boards.css.map */
