--- src/components/Pages/LoginPage/LoggedInView/Boards.scss
@@ -1,9 +1,0 @@
-/** @format */
-
-.boards-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-}
