--- src/components/Pages/LoginPage/LoggedInView/Components/BoardBox/BoardBox.scss
@@ -1,18 +1,0 @@
-/** @format */
-
-.board-box {
-  background: var(--linear-blue);
-  color: white;
-  width: 80%;
-  margin-top: 1rem;
-}
-
-.board-link {
-  color: var(--dark);
-}
-
-.board-link:hover,
-.board-link:active,
-.board-link:focus {
-  color: white;
-}
