--- src/components/Pages/LoginPage/LoggedInView/Components/ThreadBox/ThreadBox.scss
@@ -1,42 +1,0 @@
-/** @format */
-
-.thread-box {
-  background: var(--linear-blue);
-  color: white;
-  margin-top: 1rem;
-  padding: 30px;
-  border-radius: 0.25rem;
-}
-
-.thread-container {
-  background-color: #2e2e2e;
-  padding: 30px;
-  border-radius: 0.25rem;
-}
-
-.thread-box-title {
-  font-family: var(--display-font);
-  font-size: 30px;
-}
-
-.replies-h3-container {
-  margin: 30px 0;
-}
-
-.reply {
-  background: var(--blue);
-  color: black;
-  padding: 20px;
-  margin: 5px;
-  border-radius: 0.25rem;
-}
-
-#reply {
-  color: white;
-}
-
-.show-all-replies-btn {
-  margin: 30px 0;
-  color: black !important;
-  border-color: black !important;
-}
