--- src/components/Pages/LoginPage/LoggedInView/Components/Threads/Threads.css
@@ -1,27 +1,0 @@
-/** @format */
-.threads-container {
-  width: 100%;
-  min-height: 70vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 20px;
-}
-
-.add-thread-form-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.add-thread-form {
-  width: 80vw;
-}
-
-.add-thread-form-heading {
-  font-size: 20px;
-  margin: 0 30px;
-}
-
-/*# sourceMappingURL=Threads.css.map */
