--- src/components/Pages/LoginPage/LoginForm/LoginForm.css
@@ -1,12 +1,0 @@
-/** @format */
-.login-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.login-form-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
-
-/*# sourceMappingURL=LoginForm.css.map */
