--- src/components/Pages/LoginPage/LoginForm/LoginForm.scss
@@ -1,11 +1,0 @@
-/** @format */
-
-.login-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.login-form-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
