--- src/components/Pages/LoginPage/RegistrationForm/RegistrationForm.css
@@ -1,22 +1,0 @@
-/** @format */
-.registration-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.registration-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
-
-#request-activation-link {
-  color: var(--blue);
-}
-
-.emphasis {
-  background: rgba(63, 63, 63, 0.219);
-  font-weight: bold;
-  color: white;
-}
-
-/*# sourceMappingURL=RegistrationForm.css.map */
