--- src/components/Pages/LoginPage/ResetPasswordForm/ResetPasswordForm.scss
@@ -1,11 +1,0 @@
-/** @format */
-
-.reset-password-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.reset-password-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
