--- src/components/Pages/LoginPage/UserPortalNav.scss
@@ -1,50 +1,0 @@
-/** @format */
-
-.user-portal {
-  text-align: left;
-  @media (min-width: 767px) and (max-width: 991px) {
-    .nav-item {
-      i {
-        display: none;
-      }
-    }
-    .user-portal-link {
-      padding: 0;
-    }
-  }
-}
-
-label[for="hamburger2"] {
-  margin-right: 15px;
-  @media (min-width: 767px) and (max-width: 991px) {
-    display: none;
-  }
-}
-
-.user-portal-logo {
-  font-family: var(--logo-font);
-  font-size: 130%;
-  @media (max-width: 767px) {
-    margin-left: 15px;
-  }
-}
-
-.login-panel {
-  background: black;
-}
-
-@media (max-width: 767px) {
-  .login-panel {
-    line-height: 2.5;
-  }
-}
-
-.user-portal-link {
-  padding: 20px;
-  color: var(--blue);
-  &:hover,
-  &:active,
-  &:focus {
-    color: white !important;
-  }
-}
