--- src/components/Pages/LoginPage/VerificationForm/VerificationForm.css
@@ -1,12 +1,0 @@
-/** @format */
-.verification-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.verification-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
-
-/*# sourceMappingURL=VerificationForm.css.map */
