--- src/components/Pages/LoginPage/VerificationForm/VerificationForm.scss
@@ -1,11 +1,0 @@
-/** @format */
-
-.verification-header-container {
-  width: 100%;
-  text-align: center;
-}
-
-.verification-header {
-  font-family: var(--logo-font);
-  margin-top: 30px;
-}
